chatbot/requirements.txt
# Chatbot Dependencies

# API dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.8

# Chatbot libraries
google-generativeai==0.3.1
python-dotenv==1.0.0

# Utilities
requests==2.30.0 